refactor(#1082): move CefrConstants to Services namespace alongside CefrLevelNormalizer
CefrConstants is not DTO-specific — it is consumed by services and DTO attributes alike.
Moving it to Services/CefrConstants.cs matches where CefrLevelNormalizer lives. DTOs now
import using LangTeach.Api.Services.
