+++ backend/LangTeach.Api/DTOs/AssistantDtos.cs
@@ -1,3 +1,4 @@
+using LangTeach.Api.Services;
 using System.ComponentModel.DataAnnotations;
 using System.Text.Json;
