+++ backend/LangTeach.Api/DTOs/CreateLessonRequest.cs
@@ -1,3 +1,4 @@
+using LangTeach.Api.Services;
 using System.ComponentModel.DataAnnotations;
 
 namespace LangTeach.Api.DTOs;
